requirements.txt
opencv-contrib-python==4.8.0.74
opencv-python==4.7.0.72
opencv-python-headless==4.8.0.74
mediapipe==0.10.2
streamlit==1.24.0
streamlit-authenticator==0.2.2
streamlit-javascript==0.1.5
streamlit-lottie==0.0.5
PyYAML==6.0
pandas==2.0.3
hydralit-components==1.0.10